fix: correctly handling epoch looping

+++ contracts/libraries/PanopticMath.sol
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.24;
-import "forge-std/Test.sol";
 // Interfaces
 import {CollateralTracker} from "@contracts/CollateralTracker.sol";
 import {PanopticPool} from "@contracts/PanopticPool.sol";
@@ -308,7 +307,7 @@ library PanopticMath {
                 currentEpoch = (block.timestamp >> 6) & 0xFFFFFF; // mod 2**24
                 uint256 recordedEpoch = medianData >> 232;
                 differentEpoch = currentEpoch != recordedEpoch;
-                timeDelta = int256((currentEpoch - recordedEpoch) * 64);
+                timeDelta = int256(uint256(uint24(currentEpoch - recordedEpoch))) * 64;
             }
 
             // only proceed if last entry is in a different epoch (takes care of looping edge case in a way that ">" doesn't)
